Branched to fix DMA transferring twice as many words as requested
The KL25Z DMA configured in cycle-steal (one transfer per request,
according to manual) PIT-triggered mode systematically transfers twice
as much as it should, regardless of transfer size, address alignment,
address adjustment, address modulo, type of transfer (memory to
peripheral or memory to memory) and request source (always enabled or
from peripheral).

The net result is that for memory to peripheral transfers even-numbered
transfers are immediately overwritten with odd-numbered transfers. For
the DAC and PWM, this means a wave period twice as short and half as
smooth as it should be.